Add update functionality and replace Hibernate with Spring Data JPA
A PUT mapping for updating employees was added to the RESTController. Replaced Hibernate implementation with Spring Data JPA in EmployeeDAOImpl to follow better practices. This resulted in the addition of EmployeeRepository using JpaRepository. Added spring-boot-starter-actuator dependency to the pom.xml. Refactored Employee entity class to include getter and setter methods, instead of using Lombok's @Data annotation for simplicity of future updates and removing unnecessary dependency. Also, @EnableJpaRepositories annotation was added to the main application class to enable JPA repositories. This transition was made to simplify the code and remove direct dependency on Hibernate.