Remove redundant Employee DAO/Service layers
This commit removed the redundant EmployeeDAO and EmployeeService layers and their implementations, to simplify the architecture. Instead, we rely on the Spring Data JPA's JpaRepository for CRUD operations. The changes simplify the codebase by removing unnecessary layers and enable us to take advantage of Spring Data JPA's features. Also, adjustments were made to the pom.xml file for supporting the above changes.